Support import from ESM
