+++ package.json
@@ -13,7 +13,18 @@
 	"engines": {
 		"node": ">=6"
 	},
+	"type": "module",
+	"main": "./dist/cjs/index.js",
+	"module": "./dist/mjs/index.js",
+	"exports": {
+		".": {
+			"require": "./dist/cjs/index.js",
+			"import": "./dist/mjs/index.js"
+		}
+	},
+	"types": "./dist/mjs/index.d.ts",
 	"scripts": {
+		"build": "rimraf cjs/**,mjs/** && tsc -p tsconfig.cjs.json && tsc -p tsconfig.mjs.json",
 		"test": "xo && ava && tsd"
 	},
 	"files": [
@@ -38,7 +49,9 @@
 	],
 	"devDependencies": {
 		"ava": "^1.4.1",
-		"tsd": "^0.7.2",
+		"rimraf": "^3.0.2",
+		"tsd": "^0.7.4",
+		"typescript": "^4.3.5",
 		"xo": "^0.24.0"
 	}
 }
