+++ tsconfig.cjs.json
@@ -1,0 +1,8 @@
+{
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+    "module": "commonjs",
+    "outDir": "./dist/cjs",
+		"declarationDir": "./dist/cjs"
+  }
+}
