+++ tsconfig.mjs.json
@@ -1,0 +1,9 @@
+{
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+    "module": "ESNext",
+		"rootDir": "src",
+		"outDir": "./dist/mjs",
+		"declarationDir": "./dist/mjs"
+  }
+}
